internal-test-helpers: Move `NodeQuery` class out of the `test-cases` folder
