+++ packages/internal-test-helpers/lib/node-query.js
@@ -1,7 +1,7 @@
 /* global Node */
 
 import { assert } from '@ember/debug';
-import { fireEvent, focus, matches } from '../system/synthetic-events';
+import { fireEvent, focus, matches } from './system/synthetic-events';
 
 export default class NodeQuery {
   static query(selector, context = document) {
